==> Projects/SV_BA/s10.Functional_annotation.Rmd <==
---
title: "Functional annotation"
author: "Daoming Wang"
date: "2021/08/10"
output:
  html_document: 
    theme: flatly
    highlight: espresso
    toc: true
    toc_depth: 4
    toc_float: true
  word_document: default
  pdf_document:
    includes:
      in_header: header.tex
      keep_tex: yes
      latex_engine: xelatex
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## 1 Preparation

### 1.1 Import

Import packages and functions.

```{r 1.1, echo=TRUE, message=FALSE, warning=FALSE, paged.print=FALSE}
source("functions.R")
```

### 1.2 Inputs

Read input files.

```{r 1.2, echo=TRUE}
vsv_ba_lm_adjAbun_res.sig.anno.edge<-read.table("09.Microbial_GWAS/vsv_ba_lm_adjAbun.sig.anno.final.tsv", 
                                                sep = "\t", header = T, check.names = F)
dsv_ba_lm_adjAbun_res.sig.anno.edge<-read.table("09.Microbial_GWAS/dsv_ba_lm_adjAbun.sig.anno.final.tsv", 
                                                sep = "\t", header = T, check.names = F)

vsv_ba_lm_adjAbunPCs_res.sig.anno.edge<-read.table("09.Microbial_GWAS/vsv_ba_lm_adjAbunPCs.sig.anno.final.tsv", 
                                                sep = "\t", header = T, check.names = F)
dsv_ba_lm_adjAbunPCs_res.sig.anno.edge<-read.table("09.Microbial_GWAS/dsv_ba_lm_adjAbunPCs.sig.anno.final.tsv", 
                                                sep = "\t", header = T, check.names = F)

vsv_info<-read.table("01.cleanData/SV_info/20200801_LLD_300OB_vsgv_info_anno.tsv",
                     sep = "\t",header = T,stringsAsFactors = F, quote = "")
dsv_info<-read.table("01.cleanData/SV_info/20200801_LLD_300OB_dsgv_info_anno.tsv",
                     sep = "\t",header = T,stringsAsFactors = F,quote = "")

ba_gene<-read.table("00.rawData/BA_gene/Gene_list - clean.txt", sep = "\t",header = T, check.names = F)

B.wex_scaf<-read.csv("00.rawData/Single_bacteria/Blautia wexlerae DSM 19850/sequence_data_export.txt",
                     sep = "\t",header = T,stringsAsFactors = F, check.names = F)
B.wex<-read.csv("00.rawData/Single_bacteria/Blautia wexlerae DSM 19850/Blautia wexlerae DSM 19850.txt",
                sep = "\t",header = T, stringsAsFactors = F, check.names = F)

C.com_scaf<-read.csv("00.rawData/Single_bacteria/Coprococcus comes ATCC 27758/sequence_data_export.txt",
                     sep = "\t",header = T,stringsAsFactors = F, check.names = F)
C.com<-read.csv("00.rawData/Single_bacteria/Coprococcus comes ATCC 27758/Coprococcus comes ATCC 27758.txt",
                sep = "\t",header = T, stringsAsFactors = F, check.names = F)

E.ven_scaf<-read.csv("00.rawData/Single_bacteria/Eubacterium ventriosum ATCC 27560/sequence_data_export.txt",
                     sep = "\t",header = T,stringsAsFactors = F, check.names = F)
E.ven<-read.csv("00.rawData/Single_bacteria/Eubacterium ventriosum ATCC 27560/Eubacterium ventriosum ATCC 27560.txt",
                sep = "\t",header = T, stringsAsFactors = F, check.names = F)

B.mas_scaf<-read.csv("00.rawData/Single_bacteria/Bacteroides massiliensis B84634 = Timone 84634 = DSM 17679 = JCM 13223/sequence_data_export.txt",
                     sep = "\t",header = T,stringsAsFactors = F, check.names = F)
B.mas<-read.csv("00.rawData/Single_bacteria/Bacteroides massiliensis B84634 = Timone 84634 = DSM 17679 = JCM 13223/Bacteroides massiliensis B84634 Timone 84634 DSM 17679 JCM 13223 dm.txt",
                sep = "\t",header = T, stringsAsFactors = F, check.names = F)

E.hal_scaf<-read.csv("00.rawData/Single_bacteria/Bacteroides massiliensis B84634 = Timone 84634 = DSM 17679 = JCM 13223/sequence_data_export.txt",
                     sep = "\t",header = T,stringsAsFactors = F, check.names = F)
E.hal<-read.csv("00.rawData/Single_bacteria/Bacteroides massiliensis B84634 = Timone 84634 = DSM 17679 = JCM 13223/Bacteroides massiliensis B84634 Timone 84634 DSM 17679 JCM 13223 dm.txt",
                sep = "\t",header = T, stringsAsFactors = F, check.names = F)

B.xyl_scaf<-read.csv("00.rawData/Single_bacteria/Bacteroides xylanisolvens XB1A/sequence_data_export.txt",
                     sep = "\t",header = T,stringsAsFactors = F, check.names = F)
B.xyl<-read.csv("00.rawData/Single_bacteria/Bacteroides xylanisolvens XB1A/Bacteroides xylanisolvens XB1A dm.txt",
                sep = "\t",header = T, stringsAsFactors = F, check.names = F)

```


### 2 Get BA-associated SV list

```{r 2}
write.table(vsv_info$Genes, "10.Functional_annotation/all_vsv_gene.tsv", sep = "\t", row.names = F, col.names = F, quote = F)
write.table(dsv_info$Genes, "10.Functional_annotation/all_dsv_gene.tsv", sep = "\t", row.names = F, col.names = F, quote = F)

all_sv_gene<-c(vsv_info$Genes, dsv_info$Genes)
write.table(all_sv_gene, "10.Functional_annotation/all_sv_gene.tsv", sep = "\t", row.names = F, col.names = F, quote = F)

ba_assoc_vsv<-c(vsv_ba_lm_adjAbun_res.sig.anno.edge$SV,vsv_ba_lm_adjAbunPCs_res.sig.anno.edge$SV) %>% unique
ba_assoc_dsv<-c(dsv_ba_lm_adjAbun_res.sig.anno.edge$SV,dsv_ba_lm_adjAbunPCs_res.sig.anno.edge$SV) %>% unique

ba_assoc_vsv_gene<-c(vsv_ba_lm_adjAbun_res.sig.anno.edge$Genes,vsv_ba_lm_adjAbunPCs_res.sig.anno.edge$Genes) %>% unique
write.table(ba_assoc_vsv_gene, "10.Functional_annotation/ba_assoc_vsv_gene.tsv", sep = "\t", row.names = F, col.names = F, quote = F)

ba_assoc_dsv_gene<-c(dsv_ba_lm_adjAbun_res.sig.anno.edge$Genes,dsv_ba_lm_adjAbunPCs_res.sig.anno.edge$Genes) %>% unique
write.table(ba_assoc_dsv_gene, "10.Functional_annotation/ba_assoc_dsv_gene.tsv", sep = "\t", row.names = F, col.names = F, quote = F)

ba_assoc_sv_gene<-c(ba_assoc_vsv_gene, ba_assoc_dsv_gene) %>% unique
write.table(ba_assoc_sv_gene, "10.Functional_annotation/ba_assoc_sv_gene.tsv", sep = "\t", row.names = F, col.names = F, quote = F)
```


### 3 Get bile acid genes

```{r 3}
B.wex_gene<-reCalcuPos(B.wex, B.wex_scaf)

ba_gene_keyword<-c("Bile acid","Choloylglycine","sulfatase")

B.wex_gene_ba<-B.wex_gene[grep(paste(ba_gene_keyword,collapse="|"), B.wex_gene$function., ignore.case = T),]

if(!dir.exists("10.Functional_annotation")){dir.create("10.Functional_annotation")}
write.table(B.wex_gene, "10.Functional_annotation/B.wex_gene.tsv",
            col.names = T, row.names = F, sep = "\t", quote = F)
write.table(B.wex_gene_ba, "10.Functional_annotation/B.wex_gene_ba.tsv",
            col.names = T, row.names = F, sep = "\t", quote = F)

##
C.com_gene<-reCalcuPos(C.com, C.com_scaf)
E.ven_gene<-reCalcuPos(E.ven, E.ven_scaf)
B.mas_gene<-reCalcuPos(B.mas, B.mas_scaf)
E.hal_gene<-reCalcuPos(E.hal, E.hal_scaf)
B.xyl_gene<-reCalcuPos(B.xyl, B.xyl_scaf)

write.table(C.com_gene, "10.Functional_annotation/C.com_gene.tsv",
            col.names = T, row.names = F, sep = "\t", quote = F)
write.table(E.ven_gene, "10.Functional_annotation/E.ven_gene.tsv",
            col.names = T, row.names = F, sep = "\t", quote = F)
write.table(B.mas_gene, "10.Functional_annotation/B.mas_gene.tsv",
            col.names = T, row.names = F, sep = "\t", quote = F)
write.table(E.hal_gene, "10.Functional_annotation/E.hal_gene.tsv",
            col.names = T, row.names = F, sep = "\t", quote = F)
write.table(B.xyl_gene, "10.Functional_annotation/B.xyl_gene.tsv",
            col.names = T, row.names = F, sep = "\t", quote = F)
```

### 4 BA genes

```{r 4}
ba_gene_long<-gather(ba_gene, BA_gene, Feature_id, c(3:18)) %>% na.omit
ba_gene_long$Feature_id_1 <- str_extract_all(ba_gene_long$Feature_id, ",.*") %>% str_replace_all(", ", "") %>% as.numeric
ba_gene_long$Feature_id   <- ba_gene_long$Feature_id %>% str_replace_all(",.*", "") %>% as.numeric

ba_gene_long$Feature_id<-str_extract_all(ba_gene_long$`Organism (PubSEED ID)`, "\\(.+\\)") %>% str_replace_all("\\(|\\)", "") %>% paste("fig",.,sep = "|") %>% paste(.,".peg.",ba_gene_long$Feature_id,sep = "")
ba_gene_long$Feature_id_1<-str_extract_all(ba_gene_long$`Organism (PubSEED ID)`, "\\(.+\\)") %>% str_replace_all("\\(|\\)", "") %>% paste("fig",.,sep = "|") %>% paste(.,".peg.",ba_gene_long$Feature_id_1,sep = "")
ba_gene_long$Feature_id_1[grep("peg.NA",ba_gene_long$Feature_id_1)]<-NA

if(!dir.exists("10.Functional_annotation/PATRIC")){dir.create("10.Functional_annotation/PATRIC")}
write.table(ba_gene_long, "10.Functional_annotation/PATRIC/ba_gene_our_genomes.tsv",col.names = T, row.names = F, sep = "\t", quote = F)

```
